requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
Django==5.2.2
django-allauth==65.9.0
django-cors-headers==4.7.0
idna==3.10
pycparser==2.22
PyJWT==2.4.0
PyMySQL==1.1.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0

## pip freeze 명령어를 통해 확인한 패키지 목록
## 프로젝트에서 사용하는 패키지 목록입니다.
## pip install -r requirements.txt를 통해 필요 패키지 자동 설치가 가능합니다.